+++ spec/surround-with-view-spec.js
@@ -4,6 +4,6 @@ import SurroundWithView from '../lib/surround-with-view';
 
 describe('SurroundWithView', () => {
   it('has one valid test', () => {
-    expect('life').toBe('easy');
+    expect('life').toBe('life');
   });
 });